Move new client to a simpler package
