+++ oslc-client/src/main/java/org/eclipse/lyo/client/OslcClientFactory.java
@@ -1,11 +1,11 @@
-package org.eclipse.lyo.oslc4j.client;
+package org.eclipse.lyo.client;
 
 public class OslcClientFactory {
-    
+
     static public OslcOAuthClientBuilder oslcOAuthClientBuilder() {
         return new OslcOAuthClientBuilder();
     }
-    
+
     static public OslcClientBuilder oslcClientBuilder() {
         return new OslcClientBuilder();
     }
