+++ oslc-client/src/main/java/org/eclipse/lyo/client/RootServicesHelper.java
@@ -16,14 +16,13 @@
  *     Samuel Padgett      - add request consumer key and OAuth approval module URLs
  *     Samuel Padgett      - handle trailing '/' in baseUrl
  *******************************************************************************/
-package org.eclipse.lyo.oslc4j.client;
+package org.eclipse.lyo.client;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriBuilder;
