+++ oslc-client/src/main/java/org/eclipse/lyo/client/resources/OslcQueryResult.java
@@ -16,7 +16,7 @@
  *     Samuel Padgett   - support setting member property
  *     Samuel Padgett   - preserve member property on call to getNext()
  *******************************************************************************/
-package org.eclipse.lyo.oslc4j.client.resources;
+package org.eclipse.lyo.client.resources;
 
 import java.io.InputStream;
 import java.lang.reflect.InvocationTargetException;
@@ -28,7 +28,7 @@ import javax.xml.datatype.DatatypeConfigurationException;
 
 import javax.ws.rs.core.Response;
 
-import org.eclipse.lyo.oslc4j.client.OSLCConstants;
+import org.eclipse.lyo.client.OSLCConstants;
 import org.eclipse.lyo.oslc4j.core.exception.OslcCoreApplicationException;
 import org.eclipse.lyo.oslc4j.core.model.OslcConstants;
 import org.eclipse.lyo.oslc4j.provider.jena.JenaModelHelper;
